+++ game/scripts/map_generator.lua
@@ -8,11 +8,15 @@ MapGenerator.MAP_H = 1080
 MapGenerator.TUNNEL = 0
 MapGenerator.WALL = 1
 MapGenerator.BLOCKAGE = 2
+MapGenerator.SHACK = 3 
 
 -- Configuration settings
 -- Constants and config will be moved later
 MapGenerator.config = {
     mainMine = { xPosition = 0.4, width = 0.0010 },
+    shacks = {  
+        size = {width = 20, height = 16} -- Increased from 5x4 to 20x16
+    },
     levels = {
         count = {min = 4, max = 7},
         heightPercent = {min = 0.05, max = 0.2},
@@ -130,13 +134,22 @@ function MapGenerator.carveVerticalMine(map, x, yStart, yEnd)
 end
 
 -- Carve a horizontal tunnel
-function MapGenerator.carveHorizontalTunnel(map, startX, endX, y, width)
+function MapGenerator.carveHorizontalTunnel(map, startX, endX, y, width, levelIndex)
     for x = startX, endX do
         for w = 0, width - 1 do
             local yPos = y + w
             if yPos <= MapGenerator.MAP_H then
                 map[yPos][x] = MapGenerator.TUNNEL
             end
+        end
+    end
+    
+    -- Add shacks
+    if levelIndex == 1 or levelIndex % 2 == 1 then
+        local groundX = random(startX, endX)
+        local groundY = y + width - 1
+        if groundY <= MapGenerator.MAP_H then
+            map[groundY][groundX] = MapGenerator.SHACK
         end
     end
 end
@@ -554,7 +567,7 @@ function MapGenerator.generateMine()
                 right = mainX
             })
             
-            MapGenerator.carveHorizontalTunnel(map, leftEnd, mainX, y, tunnelWidth)
+            MapGenerator.carveHorizontalTunnel(map, leftEnd, mainX, y, tunnelWidth, i)
             
             -- Procces right side
             local rightLength = floor(MapGenerator.MAP_W * random(
@@ -570,7 +583,7 @@ function MapGenerator.generateMine()
                 right = rightEnd
             })
             
-            MapGenerator.carveHorizontalTunnel(map, mainX, rightEnd, y, tunnelWidth)
+            MapGenerator.carveHorizontalTunnel(map, mainX, rightEnd, y, tunnelWidth, i)
         end
     end